refactor: reorganize openapi directory structure
