+++ packages/hoppscotch-sh-admin/src/components.d.ts
@@ -33,8 +33,10 @@ declare module 'vue' {
     HoppSmartTable: typeof import('@hoppscotch/ui')['HoppSmartTable']
     HoppSmartTabs: typeof import('@hoppscotch/ui')['HoppSmartTabs']
     HoppSmartToggle: typeof import('@hoppscotch/ui')['HoppSmartToggle']
+    IconLucideAlertTriangle: typeof import('~icons/lucide/alert-triangle')['default']
     IconLucideArrowLeft: typeof import('~icons/lucide/arrow-left')['default']
     IconLucideArrowUpRight: typeof import('~icons/lucide/arrow-up-right')['default']
+    IconLucideBadgeCheck: typeof import('~icons/lucide/badge-check')['default']
     IconLucideCheck: typeof import('~icons/lucide/check')['default']
     IconLucideChevronDown: typeof import('~icons/lucide/chevron-down')['default']
     IconLucideHelpCircle: typeof import('~icons/lucide/help-circle')['default']
@@ -42,13 +44,21 @@ declare module 'vue' {
     IconLucideInfo: typeof import('~icons/lucide/info')['default']
     IconLucideSearch: typeof import('~icons/lucide/search')['default']
     IconLucideUser: typeof import('~icons/lucide/user')['default']
+    OnboardingAuthProviderCard: typeof import('./components/onboarding/AuthProviderCard.vue')['default']
+    OnboardingAuthSetup: typeof import('./components/onboarding/AuthSetup.vue')['default']
+    OnboardingCompleteScreen: typeof import('./components/onboarding/CompleteScreen.vue')['default']
+    OnboardingOAuthSetup: typeof import('./components/onboarding/OAuthSetup.vue')['default']
+    OnboardingSmtpSetup: typeof import('./components/onboarding/SmtpSetup.vue')['default']
+    OnboardingWelcomeScreen: typeof import('./components/onboarding/WelcomeScreen.vue')['default']
     SettingsAuthConfiguration: typeof import('./components/settings/AuthConfiguration.vue')['default']
     SettingsAuthConfigurations: typeof import('./components/settings/AuthConfigurations.vue')['default']
     SettingsAuthProvider: typeof import('./components/settings/AuthProvider.vue')['default']
+    SettingsAuthToken: typeof import('./components/settings/AuthToken.vue')['default']
     SettingsConfigurations: typeof import('./components/settings/Configurations.vue')['default']
     SettingsDataSharing: typeof import('./components/settings/DataSharing.vue')['default']
     SettingsHistoryConfiguration: typeof import('./components/settings/HistoryConfiguration.vue')['default']
     SettingsOAuthProviderConfigurations: typeof import('./components/settings/OAuthProviderConfigurations.vue')['default']
+    SettingsRateLimit: typeof import('./components/settings/RateLimit.vue')['default']
     SettingsReset: typeof import('./components/settings/Reset.vue')['default']
     SettingsServerRestart: typeof import('./components/settings/ServerRestart.vue')['default']
     SettingsSmtpConfiguration: typeof import('./components/settings/SmtpConfiguration.vue')['default']
@@ -63,6 +73,7 @@ declare module 'vue' {
     TokensGenerateModal: typeof import('./components/tokens/GenerateModal.vue')['default']
     TokensList: typeof import('./components/tokens/List.vue')['default']
     TokensOverview: typeof import('./components/tokens/Overview.vue')['default']
+    UiAccordion: typeof import('./components/ui/Accordion.vue')['default']
     UiAutoResetIcon: typeof import('./components/ui/AutoResetIcon.vue')['default']
     UsersDetails: typeof import('./components/users/Details.vue')['default']
     UsersInviteModal: typeof import('./components/users/InviteModal.vue')['default']
